Implemented the loans aspect (OneToMany and ManyToOne). Every

+++ ImoLandRegistry/src/main/java/com/imo/landregistry/dao/OwnerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.imo.landregistry.entity.OwnerEntity;
 
-public interface OwnerRepository extends JpaRepository<OwnerEntity, Long> {
+public interface OwnerRepository extends JpaRepository<OwnerEntity, Integer> {
 
 }
